Move page-related column definitions to wagtail.admin.ui.tables.pages
