+++ wagtail/admin/views/pages/usage.py
@@ -6,32 +6,15 @@ from django.http import Http404
 from django.urls import reverse
 from django.utils.translation import gettext_lazy as _
 
-from wagtail.admin.ui.tables import BaseColumn, Column, DateColumn
+from wagtail.admin.ui.tables import Column, DateColumn
+from wagtail.admin.ui.tables.pages import (
+    PageStatusColumn,
+    PageTitleColumn,
+    ParentPageColumn,
+)
 from wagtail.admin.views import generic
 from wagtail.admin.views.generic.base import BaseListingView
 from wagtail.models import Page
-
-
-class PageTitleColumn(BaseColumn):
-    cell_template_name = "wagtailadmin/pages/listing/_page_title_cell.html"
-
-    def get_cell_context_data(self, instance, parent_context):
-        context = super().get_cell_context_data(instance, parent_context)
-        context["page_perms"] = instance.permissions_for_user(
-            parent_context["request"].user
-        )
-        return context
-
-
-class ParentPageColumn(Column):
-    cell_template_name = "wagtailadmin/pages/listing/_parent_page_cell.html"
-
-    def get_value(self, instance):
-        return instance.get_parent()
-
-
-class PageStatusColumn(BaseColumn):
-    cell_template_name = "wagtailadmin/pages/listing/_page_status_cell.html"
 
 
 class ContentTypeUseView(BaseListingView):
